improve dagster import perf [INTERNAL_BRANCH=al/08-15-fix_import]
