+++ integration_tests/test_suites/daemon-test-suite/conftest.py
@@ -4,7 +4,7 @@ import pytest
 
 from dagster import file_relative_path
 from dagster._core.test_utils import instance_for_test
-from dagster._core.workspace import WorkspaceProcessContext
+from dagster._core.workspace.context import WorkspaceProcessContext
 from dagster._core.workspace.load_target import PythonFileTarget
