+++ python_modules/dagit/dagit/debug.py
@@ -5,7 +5,7 @@ import click
 from dagster import DagsterInstance
 from dagster import _check as check
 from dagster._cli.debug import DebugRunPayload
-from dagster._core.workspace import WorkspaceProcessContext
+from dagster._core.workspace.context import WorkspaceProcessContext
 from dagster._serdes import deserialize_json_to_dagster_namedtuple
 
 from .cli import (
