+++ python_modules/dagster-graphql/dagster_graphql/test/utils.py
@@ -6,7 +6,7 @@ from graphql import graphql
 import dagster._check as check
 from dagster._core.instance import DagsterInstance
 from dagster._core.test_utils import wait_for_runs_to_finish
-from dagster._core.workspace import WorkspaceProcessContext
+from dagster._core.workspace.context import WorkspaceProcessContext
 from dagster._core.workspace.load_target import PythonFileTarget
