+++ python_modules/dagster/dagster/_core/storage/event_log/__init__.py
@@ -1,11 +1,4 @@
-from .base import (
-    AssetRecord,
-    EventLogEntry,
-    EventLogRecord,
-    EventLogStorage,
-    EventRecordsFilter,
-    RunShardedEventsCursor,
-)
+from .base import AssetRecord, EventLogRecord, EventLogStorage
 from .in_memory import InMemoryEventLogStorage
 from .polling_event_watcher import SqlPollingEventWatcher
 from .schema import AssetKeyTable, SqlEventLogStorageMetadata, SqlEventLogStorageTable
