+++ python_modules/dagster/dagster/_core/telemetry.py
@@ -466,7 +466,7 @@ def log_repo_stats(instance, source, pipeline=None, repo=None):
 
 
 def log_workspace_stats(instance, workspace_process_context):
-    from dagster._core.workspace import IWorkspaceProcessContext
+    from dagster._core.workspace.context import IWorkspaceProcessContext
 
     check.inst_param(instance, "instance", DagsterInstance)
     check.inst_param(
