+++ python_modules/dagster/dagster/_daemon/backfill.py
@@ -12,7 +12,7 @@ from dagster._core.execution.backfill import (
 from dagster._core.instance import DagsterInstance
 from dagster._core.storage.pipeline_run import PipelineRun, RunsFilter
 from dagster._core.storage.tags import PARTITION_NAME_TAG
-from dagster._core.workspace import IWorkspace
+from dagster._core.workspace.context import IWorkspace
 from dagster._utils.error import serializable_error_info_from_exc_info
 
 # out of abundance of caution, sleep at checkpoints in case we are pinning CPU by submitting lots
