+++ python_modules/dagster/dagster/_daemon/sensor.py
@@ -28,7 +28,7 @@ from dagster._core.scheduler.instigation import (
 from dagster._core.storage.pipeline_run import PipelineRun, PipelineRunStatus, RunsFilter
 from dagster._core.storage.tags import RUN_KEY_TAG, SENSOR_NAME_TAG
 from dagster._core.telemetry import SENSOR_RUN_CREATED, hash_name, log_action
-from dagster._core.workspace import IWorkspace
+from dagster._core.workspace.context import IWorkspace
 from dagster._utils import merge_dicts
 from dagster._utils.error import serializable_error_info_from_exc_info
