+++ python_modules/dagster/dagster/_scheduler/scheduler.py
@@ -26,7 +26,7 @@ from dagster._core.scheduler.scheduler import DEFAULT_MAX_CATCHUP_RUNS, DagsterS
 from dagster._core.storage.pipeline_run import PipelineRun, PipelineRunStatus, RunsFilter
 from dagster._core.storage.tags import RUN_KEY_TAG, SCHEDULED_EXECUTION_TIME_TAG
 from dagster._core.telemetry import SCHEDULED_RUN_CREATED, hash_name, log_action
-from dagster._core.workspace import IWorkspace
+from dagster._core.workspace.context import IWorkspace
 from dagster._seven.compat.pendulum import to_timezone
 from dagster._utils import merge_dicts
 from dagster._utils.error import serializable_error_info_from_exc_info
