+++ python_modules/dagster/dagster_tests/cli_tests/workspace_tests/multi_location/test_multi_location_workspace.py
@@ -7,7 +7,7 @@ import yaml
 from dagster import DagsterInstance
 from dagster._core.host_representation import GrpcServerRepositoryLocation
 from dagster._core.test_utils import instance_for_test
-from dagster._core.workspace import WorkspaceProcessContext
+from dagster._core.workspace.context import WorkspaceProcessContext
 from dagster._core.workspace.load import (
     load_workspace_process_context_from_yaml_paths,
     location_origins_from_config,
