+++ python_modules/libraries/dagster-shell/dagster_shell_tests/test_terminate.py
@@ -7,7 +7,7 @@ from dagster_shell.utils import execute
 from dagster import repository
 from dagster._core.storage.pipeline_run import PipelineRunStatus
 from dagster._core.test_utils import instance_for_test, poll_for_finished_run, poll_for_step_start
-from dagster._core.workspace import WorkspaceProcessContext
+from dagster._core.workspace.context import WorkspaceProcessContext
 from dagster._core.workspace.load_target import PythonFileTarget
 from dagster._legacy import pipeline, solid
 from dagster._utils import file_relative_path
